keeper-commander: fix clipboard command
keeper-commander relies on python3-pyperclip to copy text into the
clipboard. This in turn does a lot of detection of what the system
supports. Upstream pyperclip work fine, but Ubuntu's package patches it
to use gi, which doesn't work.

So, patch keeper to instead directly call xclip in a subprocess